+++ crates/scheduler/src/event/platform.rs
@@ -1,0 +1,53 @@
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use wasefire_board_api::platform::Event;
+use wasefire_board_api::Api as Board;
+
+#[cfg(all(feature = "board-api-platform-protocol", feature = "applet-api-platform-protocol"))]
+pub mod protocol;
+
+#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub enum Key {
+    #[cfg(all(feature = "board-api-platform-protocol", feature = "applet-api-platform-protocol"))]
+    Protocol(protocol::Key),
+}
+
+impl<B: Board> From<Key> for crate::event::Key<B> {
+    fn from(key: Key) -> Self {
+        crate::event::Key::Platform(key)
+    }
+}
+
+impl<'a> From<&'a Event> for Key {
+    fn from(event: &'a Event) -> Self {
+        match *event {
+            #[cfg(all(
+                feature = "board-api-platform-protocol",
+                feature = "applet-api-platform-protocol"
+            ))]
+            Event::Protocol(ref event) => Key::Protocol(event.into()),
+        }
+    }
+}
+
+pub fn process(event: Event) {
+    match event {
+        #[cfg(all(
+            feature = "board-api-platform-protocol",
+            feature = "applet-api-platform-protocol"
+        ))]
+        Event::Protocol(_) => protocol::process(),
+    }
+}