+++ crates/api-macro/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.6.1-git
 
+### Minor
+
+- Support the `api-platform-protocol` feature
+
 ### Patch
 
 - Return a user error instead of an internal error for enums `TryFrom`